+++ NEWS.md
@@ -14,6 +14,7 @@
 ## Bug fixes
 
 - Fixed issue in `model_parameters()` when `robust = TRUE`, which could sometimes mess up order of the statistic column.
+- Fixed issues in `model_parameters()` with wrong `df` for `lme`-models.
 - Fixed issues in `model_parameters.merMod()` when `df_method` was not set to default.
 - Fixed issues in `model_parameters.merMod()` when `robust = TRUE`.